refactor(web): migrate Chat, Agents, and AgentDetail pages to app router structure
- Moved ChatPage.tsx to app/chat/page.tsx
- Moved AgentsPage.tsx to app/agents/page.tsx
- Moved AgentDetailPage.tsx to app/agents/[id]/page.tsx
- Updated routing references in main.tsx
- Updated test imports
