+++ web/src/__tests__/AgentDelete.test.tsx
@@ -81,7 +81,7 @@ vi.mock('@/components/AgentStatusBadge', () => ({
 
 // ── Lazy import after mocks ────────────────────────────────────────────────────
 
-import AgentsPage from '@/pages/AgentsPage';
+import AgentsPage from '@/app/agents/page';
 import { toast } from 'sonner';
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
