+++ web/src/app/agents/[id]/page.tsx
